Move common document logic into one class
